+++ src/onegov.election_day/onegov/election_day/theme/election_day_theme.py
@@ -7,7 +7,7 @@ class ElectionDayTheme(BaseTheme):
 
     # don't touch this number, it's incremented using bumpversion, so every
     # release will automatically trigger a rebuild of the theme
-    version = '1.2.0'
+    version = '1.2.1'
 
     @property
     def post_imports(self):